refactor: Удалить лишние аннотации. Добавить индексы в БД.
