+++ main-service/src/main/java/ru/practicum/controller/adminPart/AdminUserController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import ru.practicum.dto.NewUserDto;
+import ru.practicum.dto.user.NewUserDto;
 import ru.practicum.service.UserService;
 
 import javax.validation.Valid;
