+++ main-service/src/main/java/ru/practicum/controller/publicPart/PublicCompilationController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import ru.practicum.dto.CompilationDto;
+import ru.practicum.dto.compilation.CompilationDto;
 import ru.practicum.service.CompilationService;
 
 import java.util.List;
@@ -17,7 +17,6 @@ import java.util.List;
 @RequestMapping(path = "/compilations")
 @RequiredArgsConstructor
 public class PublicCompilationController {
-
     private final CompilationService service;
 
     @GetMapping("/{compId}")
