--- main-service/src/main/java/ru/practicum/dto/EventFullDto.java
@@ -1,27 +1,0 @@
-package ru.practicum.dto;
-
-import lombok.Builder;
-import lombok.Data;
-import ru.practicum.model.Location;
-import ru.practicum.model.State;
-
-@Data
-@Builder
-public class EventFullDto {
-    private Long id;
-    private String annotation;
-    private NewCategoryDto category;
-    private String description;
-    private Long confirmedRequests;
-    private String createdOn;
-    private String eventDate;
-    private UserShortDto initiator;
-    private Location location;
-    private Boolean paid;
-    private Long participantLimit;
-    private String publishedOn;
-    private Boolean requestModeration;
-    private State state;
-    private String title;
-    private Long views;
-}
