+++ main-service/src/main/java/ru/practicum/dto/category/NewCategoryDto.java
@@ -1,12 +1,12 @@
-package ru.practicum.dto;
+package ru.practicum.dto.category;
 
 import lombok.Builder;
-import lombok.Data;
+import lombok.Getter;
 import org.hibernate.validator.constraints.Length;
 
 import javax.validation.constraints.NotBlank;
 
-@Data
+@Getter
 @Builder
 public class NewCategoryDto {
     private Long id;
