+++ main-service/src/main/java/ru/practicum/dto/compilation/CompilationDto.java
@@ -1,13 +1,14 @@
-package ru.practicum.dto;
+package ru.practicum.dto.compilation;
 
 import lombok.Builder;
-import lombok.Data;
+import lombok.Getter;
 import org.hibernate.validator.constraints.Length;
+import ru.practicum.dto.event.EventsShortDto;
 
 import javax.validation.constraints.NotBlank;
 import java.util.List;
 
-@Data
+@Getter
 @Builder
 public class CompilationDto {
     private Long id;
