+++ main-service/src/main/java/ru/practicum/dto/compilation/NewCompilationDto.java
@@ -1,13 +1,15 @@
-package ru.practicum.dto;
+package ru.practicum.dto.compilation;
 
 import lombok.Builder;
-import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
 import org.hibernate.validator.constraints.Length;
 
 import javax.validation.constraints.NotBlank;
 import java.util.List;
 
-@Data
+@Getter
+@Setter
 @Builder
 public class NewCompilationDto {
     private List<Long> events;
