+++ main-service/src/main/java/ru/practicum/dto/event/EventFullDto.java
@@ -1,0 +1,46 @@
+package ru.practicum.dto.event;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+import org.hibernate.validator.constraints.Length;
+import ru.practicum.dto.category.NewCategoryDto;
+import ru.practicum.dto.user.UserShortDto;
+import ru.practicum.model.Location;
+import ru.practicum.model.State;
+
+import javax.validation.constraints.NotNull;
+import java.time.LocalDateTime;
+
+import static ru.practicum.utility.UtilityClass.pattern;
+
+@Getter
+@Setter
+@Builder
+public class EventFullDto {
+    private Long id;
+    private String annotation;
+    private NewCategoryDto category;
+    private String description;
+    private Long confirmedRequests;
+    @JsonFormat
+            (shape = JsonFormat.Shape.STRING, pattern = pattern)
+    private LocalDateTime createdOn;
+    @JsonFormat
+            (shape = JsonFormat.Shape.STRING, pattern = pattern)
+    private LocalDateTime eventDate;
+    private UserShortDto initiator;
+    private Location location;
+    private Boolean paid;
+    private Long participantLimit;
+    @JsonFormat
+            (shape = JsonFormat.Shape.STRING, pattern = pattern)
+    private LocalDateTime publishedOn;
+    private Boolean requestModeration;
+    private State state;
+    @NotNull
+    @Length(min = 3, max = 120)
+    private String title;
+    private Long views;
+}
