+++ main-service/src/main/java/ru/practicum/dto/event/EventsShortDto.java
@@ -1,0 +1,27 @@
+package ru.practicum.dto.event;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.Builder;
+import lombok.Getter;
+import ru.practicum.dto.category.NewCategoryDto;
+import ru.practicum.dto.user.UserShortDto;
+
+import java.time.LocalDateTime;
+
+import static ru.practicum.utility.UtilityClass.pattern;
+
+@Getter
+@Builder
+public class EventsShortDto {
+    private Long id;
+    private String annotation;
+    private NewCategoryDto category;
+    private Long confirmedRequests;
+    @JsonFormat
+            (shape = JsonFormat.Shape.STRING, pattern = pattern)
+    private LocalDateTime eventDate;
+    private UserShortDto initiator;
+    private Boolean paid;
+    private String title;
+    private Long views;
+}
