+++ main-service/src/main/java/ru/practicum/dto/event/NewEventDto.java
@@ -1,15 +1,25 @@
-package ru.practicum.dto;
+package ru.practicum.dto.event;
 
 
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.hibernate.validator.constraints.Length;
 import ru.practicum.model.Location;
 
 import javax.validation.constraints.NotNull;
+import java.time.LocalDateTime;
 
-@Data
+import static ru.practicum.utility.UtilityClass.pattern;
+
+@Getter
+@Setter
 @Builder
+@NoArgsConstructor
+@AllArgsConstructor
 public class NewEventDto {
     @NotNull
     @Length(min = 20, max = 2000)
@@ -20,7 +30,9 @@ public class NewEventDto {
     @Length(min = 20, max = 7000)
     private String description;
     @NotNull
-    private String eventDate;
+    @JsonFormat
+            (shape = JsonFormat.Shape.STRING, pattern = pattern)
+    private LocalDateTime eventDate;
     @NotNull
     private Location location;
     private Boolean paid;
