+++ main-service/src/main/java/ru/practicum/dto/request/EventRequestStatusUpdateResult.java
@@ -1,11 +1,11 @@
-package ru.practicum.dto;
+package ru.practicum.dto.request;
 
 import lombok.Builder;
-import lombok.Data;
+import lombok.Getter;
 
 import java.util.List;
 
-@Data
+@Getter
 @Builder
 public class EventRequestStatusUpdateResult {
     private List<ParticipationRequestDto> confirmedRequests;
