+++ main-service/src/main/java/ru/practicum/dto/user/UserShortDto.java
@@ -1,9 +1,9 @@
-package ru.practicum.dto;
+package ru.practicum.dto.user;
 
 import lombok.Builder;
-import lombok.Data;
+import lombok.Getter;
 
-@Data
+@Getter
 @Builder
 public class UserShortDto {
     private Long id;
