+++ main-service/src/main/java/ru/practicum/repository/UserRepository.java
@@ -2,7 +2,6 @@ package ru.practicum.repository;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import ru.practicum.model.User;
 
 import java.util.List;
@@ -11,6 +10,4 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
     List<User> findByIdIn(List<Long> ids, Pageable pageable);
 
-    @Query("select u from User as u")
-    List<User> findAllUsers(Pageable pageable);
 }
