+++ main-service/src/main/java/ru/practicum/service/CategoryService.java
@@ -1,6 +1,6 @@
 package ru.practicum.service;
 
-import ru.practicum.dto.NewCategoryDto;
+import ru.practicum.dto.category.NewCategoryDto;
 
 import java.util.List;
