+++ main-service/src/main/java/ru/practicum/service/CompilationService.java
@@ -1,8 +1,8 @@
 package ru.practicum.service;
 
-import ru.practicum.dto.CompilationDto;
-import ru.practicum.dto.NewCompilationDto;
-import ru.practicum.dto.UpdateCompilationRequest;
+import ru.practicum.dto.compilation.CompilationDto;
+import ru.practicum.dto.compilation.NewCompilationDto;
+import ru.practicum.dto.compilation.UpdateCompilationRequest;
 
 import java.util.List;
