+++ main-service/src/main/java/ru/practicum/service/CompilationServiceImpl.java
@@ -4,10 +4,10 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.dto.CompilationDto;
-import ru.practicum.dto.EventsShortDto;
-import ru.practicum.dto.NewCompilationDto;
-import ru.practicum.dto.UpdateCompilationRequest;
+import ru.practicum.dto.compilation.CompilationDto;
+import ru.practicum.dto.compilation.NewCompilationDto;
+import ru.practicum.dto.compilation.UpdateCompilationRequest;
+import ru.practicum.dto.event.EventsShortDto;
 import ru.practicum.exception.IdNotFoundException;
 import ru.practicum.mapper.MapperCompilation;
 import ru.practicum.mapper.MapperEvent;
