+++ main-service/src/main/java/ru/practicum/service/RequestService.java
@@ -1,6 +1,6 @@
 package ru.practicum.service;
 
-import ru.practicum.dto.ParticipationRequestDto;
+import ru.practicum.dto.request.ParticipationRequestDto;
 
 import java.util.List;
