+++ main-service/src/main/java/ru/practicum/service/UserService.java
@@ -1,6 +1,6 @@
 package ru.practicum.service;
 
-import ru.practicum.dto.NewUserDto;
+import ru.practicum.dto.user.NewUserDto;
 
 import java.util.List;
