+++ stats/stats-dto/src/main/java/ru/practicum/dto/ViewStatsDto.java
@@ -2,10 +2,10 @@ package ru.practicum.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Data;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Data
+@Getter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
