+++ stats/stats-server/src/main/java/ru/practicum/service/StatsServiceImpl.java
@@ -1,7 +1,6 @@
 package ru.practicum.service;
 
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.dto.EndpointHitDto;
@@ -18,7 +17,6 @@ import java.util.stream.Collectors;
 import static ru.practicum.model.MapperDto.toEndpointHit;
 import static ru.practicum.model.MapperDto.toEndpointHitDto;
 
-@Slf4j
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
